修复编译问题

+++ FoundationKit/FoundationKit/experimental/FileHandle.cpp
@@ -22,27 +22,27 @@ int64 FileHandle::Tell()
 
 bool FileHandle::Seek(int64 NewPosition)
 {
-    return fseek(NativeFileHandle, NewPosition, SEEK_SET) == NewPosition;
+    return fseek(NativeFileHandle, static_cast<long>(NewPosition), SEEK_SET) == NewPosition;
 }
 
 bool FileHandle::SeekFromEnd(int64 NewPositionRelativeToEnd /*= 0*/)
 {
-    return fseek(NativeFileHandle, NewPositionRelativeToEnd, SEEK_END)== NewPositionRelativeToEnd;
+    return fseek(NativeFileHandle, static_cast<long>(NewPositionRelativeToEnd), SEEK_END)== NewPositionRelativeToEnd;
 }
 
 bool FileHandle::SeekFromCur(int64 NewPositionRelativeToCur /*= 0*/)
 {
-    return fseek(NativeFileHandle, NewPositionRelativeToCur, SEEK_CUR)== NewPositionRelativeToCur;
+    return fseek(NativeFileHandle, static_cast<long>(NewPositionRelativeToCur), SEEK_CUR)== NewPositionRelativeToCur;
 }
 
 bool FileHandle::Read(uint8* Destination, int64 BytesToRead)
 {
-    return fread(Destination, 1, BytesToRead, NativeFileHandle) == BytesToRead;
+    return fread(Destination, 1, static_cast<size_t>(BytesToRead), NativeFileHandle) == BytesToRead;
 }
 
 bool FileHandle::Write(const uint8* Source, int64 BytesToWrite)
 {
-    return fwrite(Source, 1, BytesToWrite, NativeFileHandle) == BytesToWrite;
+    return fwrite(Source, 1, static_cast<size_t>(BytesToWrite), NativeFileHandle) == BytesToWrite;
 }
 
 void FileHandle::Flush()
